backend/requirements.txt
flask
flask-cors
numpy
scikit-learn
matplotlib
pillow